v 1.6.6
May 27, 2016
---------------------------------
> added import command for local files too [from shell]
	import web
	import local
	import repo
> serialize at ontology level should serialize the whole thing

+++ ontospy/__init__.py
@@ -15,6 +15,7 @@ from colorama import Fore, Style
 from .core.graph import Graph, SparqlEndpoint 
 # this allows to load with 'import ontospy', and using 'ontospy.Graph' 
 
+from .core.util import printDebug
 
 
 
@@ -38,8 +39,10 @@ ONTOSPY_LOCAL_VIZ = ONTOSPY_LOCAL + "/viz"
 ONTOSPY_LOCAL_CACHE = ONTOSPY_LOCAL + "/.cache/" + VERSION + "/"
 
 ONTOSPY_LIBRARY_DEFAULT = ONTOSPY_LOCAL + "/models/"
-# ONTOSPY_LIBRARY_DEFAULT = 
-# os.path.join(os.path.expanduser('~'), 'ontospy-library')
+
+GLOBAL_DISABLE_CACHE = False  # set to True for testing
+
+
 
 
 BOOTSTRAP_ONTOLOGIES = [
@@ -148,7 +151,9 @@ def get_localontologies():
 def get_pickled_ontology(filename):
 	""" try to retrieve a cached ontology """
 	pickledfile = ONTOSPY_LOCAL_CACHE + "/" + filename + ".pickle"
-	if os.path.isfile(pickledfile):
+	if GLOBAL_DISABLE_CACHE:
+		printDebug("WARNING: DEMO MODE cache has been disabled in __init__.py ==============", "red")
+	if os.path.isfile(pickledfile) and not GLOBAL_DISABLE_CACHE:
 		try:
 			return cPickle.load(open(pickledfile, "rb"))
 		except:
@@ -161,7 +166,7 @@ def get_pickled_ontology(filename):
 def del_pickled_ontology(filename):
 	""" try to remove a cached ontology """
 	pickledfile = ONTOSPY_LOCAL_CACHE + "/" + filename + ".pickle"
-	if os.path.isfile(pickledfile):
+	if os.path.isfile(pickledfile) and not GLOBAL_DISABLE_CACHE:
 		os.remove(pickledfile)
 		return True
 	else:
@@ -172,7 +177,7 @@ def rename_pickled_ontology(filename, newname):
 	""" try to rename a cached ontology """
 	pickledfile = ONTOSPY_LOCAL_CACHE + "/" + filename + ".pickle"
 	newpickledfile = ONTOSPY_LOCAL_CACHE + "/" + newname + ".pickle"
-	if os.path.isfile(pickledfile):
+	if os.path.isfile(pickledfile) and not GLOBAL_DISABLE_CACHE:
 		os.rename(pickledfile, newpickledfile)
 		return True
 	else:
@@ -191,22 +196,23 @@ def do_pickle_ontology(filename, g=None):
 	if not g:
 		g = Graph(ONTOSPY_LOCAL_MODELS + "/" + filename)	
 	
-	try:				
-		cPickle.dump(g, open(pickledpath, "wb"))
-		# print Style.DIM + ".. cached <%s>" % pickledpath + Style.RESET_ALL
-	except Exception, e: 
-		print Style.DIM + "\n.. Failed caching <%s>" % filename + Style.RESET_ALL
-		print str(e)
-		print Style.DIM + "\n... attempting to increase the recursion limit from %d to %d" % (sys.getrecursionlimit(), sys.getrecursionlimit()*10) + Style.RESET_ALL
-
-		try:
-			sys.setrecursionlimit(sys.getrecursionlimit()*10)
+	if not GLOBAL_DISABLE_CACHE:
+		try:				
 			cPickle.dump(g, open(pickledpath, "wb"))
-			print Style.BRIGHT + "... SUCCESSFULLY cached <%s>" % pickledpath + Style.RESET_ALL
+			# print Style.DIM + ".. cached <%s>" % pickledpath + Style.RESET_ALL
 		except Exception, e: 
-			print Style.BRIGHT + "\n... Failed caching <%s>... aborting..." % filename + Style.RESET_ALL
-			print str(e)	
-		sys.setrecursionlimit(sys.getrecursionlimit()/10)
+			print Style.DIM + "\n.. Failed caching <%s>" % filename + Style.RESET_ALL
+			print str(e)
+			print Style.DIM + "\n... attempting to increase the recursion limit from %d to %d" % (sys.getrecursionlimit(), sys.getrecursionlimit()*10) + Style.RESET_ALL
+
+			try:
+				sys.setrecursionlimit(sys.getrecursionlimit()*10)
+				cPickle.dump(g, open(pickledpath, "wb"))
+				print Style.BRIGHT + "... SUCCESSFULLY cached <%s>" % pickledpath + Style.RESET_ALL
+			except Exception, e: 
+				print Style.BRIGHT + "\n... Failed caching <%s>... aborting..." % filename + Style.RESET_ALL
+				print str(e)	
+			sys.setrecursionlimit(sys.getrecursionlimit()/10)
 	return g
